First draft of a Maple-side disintegrator.  Main entry point is "disint", which takes an LO and the name of a free variable.  It only handles two cases of the syntax of LOs, namely Int/int and applyintegrand.
Also in this check in is a significant upgrade to the simplifications that
'indicator' will do.  Now things like
  -x < 5
  -x <= 10
  3 < 2 - 5*x
  x < infinity
  -infinity < x
  3 < 1/x
  3 <= 1/x
should all 'simplify' correctly to patterns that later passes will recognize
and incorporate into integration bounds.

The current design uses a queue of actions to perform.  The 3 implemented
actions are:
1. reconstruct (for rebuilding an expression)
2. change_var (for actually doing a change of variable)
3. promote (for push an integral up)

The heart of what is different is in the function 'invert', which is called when
we encounter an 'applyintegrand'.  This figures out what change of variable to
try, and calls "solve" to do some of the hard work.  Dealing with integration
bounds is the principal difficulty.