Make unproduct succeed more often using ideas from disintegration(!!)
unproduct is how unintegrate produces Plate.

unproduct means recognizing a weight expression as a product, much as
disintegration means recognizing a measure expression as a monadic bindx.

With this commit, we are making progress towards recognizing the conjugacy
between rolled Dirichlet and multinomial.  Now fromLO produces a Plate of
unrecognized measures rather than an unrecognized Plate of measures.