Make our eval sensitive to shadowing, even for Not(name)=anything
For example,
    eval('sum(piecewise(0 <= g[k], 1, 0), k = 0 .. n-1)',
         'piecewise(0 <= g[k], 1, 0) = FROZEN')
now returns the first argument unchanged, because the variable k is
shadowed.

Because evalindets calls eval, this change makes evalindets more
different from subsindets, so I changed some calls to evalindets into
calls to subsindets as appropriate.  (The call to evalindets from
IntegrationTools:-Expand should be a call to subsindets but remains
unchanged.)

The reason to make this change is to work around this bug in `simplify/piecewise/subs`:
    > simplify(Product(sum(f(x0), x0 = 0 .. m-1)+j,
                       j = 0 .. sum(piecewise(0 <= g[k], 1, 0), k = 0 .. n-1)-1));
    Product(sum(f(k),k = 0 .. m-1)+j,j = 0 .. n*piecewise(g[k] < 0,0,0 <= g[k],1)-1)