Add a big hack to simplify dice_predict.hk, gmm_gibbs.hk, naive_bayes_gibbs.hk
Due to integrating out the Dirichlet priors in these models, we end up
with ratios of the form

  B(a[1]+b[1],...,a[n]+b[n])/B(a[1],...,a[n])

in which the generalized Beta function B is defined by

  B(a[1],...,a[n])
  = GAMMA(a[1])*...*GAMMA(a[n]) / GAMMA(a[1]+...+a[n])
  = Beta(a[1],a[2]+...+a[n]) * Beta(a[2],a[3]+...+a[n]) *...* Beta(a[n-1],a[n])

and b1,...,bn are counts (usually 1 or 0).  All the Beta and GAMMA in
such a ratio should be simplified away.  This commit adds a hack_Beta
function to the simplifier, demonstrates the desired output (saved in
the worksheet fun-fact.mw), and calls for a better implementation.  At
first glance our output seems to match

  https://lingpipe-blog.com/2010/07/13/collapsed-gibbs-sampling-for-lda-bayesian-naive-bayes/