A way to express Dirichlet that avoids conditions on indices
This Hakaru code is easy for our Maple code to simplify, because it
doesn't involve any "piecewise".  But it burdens the user with treating
one final index separately (hence "finalsep" in the file name),
including in the input to the program (witness the extra argument "n" in
this file).  And we're not sure how to represent the categorical
distribution in a way that's parameterized by pair(array(prob),prob).
But anyway, here is a piece of thought-provoking Hakaru code.