Add a second call to rel_coulditbe inside KB:-assert.
We need to do this at the beginning of `assert'; if the condition is very
apparently false in the context, passing it to many Maple solvers will throw
errors. We need to do this at the end; sometimes `rel_coulditbe' can only
discover the condition is false after the simplifications performed by `assert'
have been done.

This is needed for e.g.

assert( And(exp(-(1/50)*x25^2)*exp((1/50)*x3^2) <= 1,
        1 < exp(-(1/50)*x25^2)*exp((1/50)*x3^2)),
        KB(Introduce(x25, AlmostEveryReal()), Introduce(x3, HReal())) )

in this call, both the conjuncts will be simplified by `try_improve_exp', but in
the recursive call to `assert_deny', one condition has not yet had
`try_improve_exp' applied to it when the first call to `rel_coulditbe' occurs.