Change the method of using `assert' to simplify conditions.
Instead of asserting the condition into an empty KB and extracting all resulting
conditions from the KB, assert it into the current context and subtract the
current context from that.

The result of `subtract' could also be examined for multiple `assert's
(the conjunction of those conditions could be taken) but I've not yet seen a
place where that is needed.