After reduce_Integrals_into, try to collect weights.
When `elim_intsum' gets a chance to do its thing, it recursively calls
`reduce'. But it only gets a chance to do so on individual summands of a sum of
integrals, not on the entire sum. If multiple such `elim_intsum's succeed, the
resulting expression will contain a sum with a common factor. This common factor
is not necessarily an `applyintegrand' but that is the only case I've seen.

Perhaps the need for this indicates that each constructor of DomShape should
indeed have its own callback. This was discussed way back when but rejected for
being unnecessary.

This is relevant to t63 and t60.