Fix Vector "index out of bounds" error in printSourceSpan
The uncaught exception was easy to trigger:

$ echo 'foo' | pretty -
pretty: ./Data/Vector/Generic.hs:235 ((!)): index out of bounds (1,1)
CallStack (from HasCallStack):
  error, called at ./Data/Vector/Internal/Check.hs:87:5 in vector-0.11.0.0-LMwQhhnXj8U3T5Bm1JFxG:Data.Vector.Internal.Check

The cause of the uncaught exception is that the source span of the error
stops at the end of the input, which is line 2 column 1 even though the
input only has one line.  (In other words, the EOF is at a nonexistent
location past all the characters in the input.)

While I'm at it, I made the message pinpoint multiline source spans more
informatively.