Avoid infinite loop in `simplify'
Strangely, the problematic call returns almost instantly when it is performed in
isolation (i.e. the infinite loop only occurs when it is called from some other
function, not in a Maple repl). The use of `'int'' instead of `int' doesn't seem
to affect anything.

This occurs in RoundTrip/mcmcPriopProp