Sometimes (but very rarely) SemiAlgebraic produces a very trivial output
But the output is not trivially the same as the input, in which case the final
result just gets mangled horribly. In this case, just keep the input;
if the result is trivial, the system is trivial.

There may a way to fail earlier for trivial systems, but I haven't found it;
everything else I've tried catches cases we actually do want to improve.