When there are no Partitions, keep the simplified output anyways.
It isn't even possible there are no Partitions in the input, or else the
previous case would have fired. If there are no Partitions after applying
`Simpl' to each operand, that means that every Partition in the input got
eliminated by `remove_false_pieces'. This is a simplification we really would
like to keep.