Try to teach the simplifier to perform Kronecker expansion
Per discussion at https://github.com/hakaru-dev/hakaru/issues/104 ,
the pivot equality condition may occur anywhere within the sum body,
not just in a top-level piecewise.  So we do it when we see the sum,
not when we see the piecewise (which we may never notice).

This rewrite ought to supercede the
"# Reduce product(piecewise(i=3,f(i),1),i=1..10) to f(3)"
in NewSLO/Piecewise.mpl , so we remove that.